# src/shared/infra/presentation/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"MovieRepositoryDTO": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"synopsis": {
						"type": "string"
					},
					"rating": {
						"type": "string"
					},
					"available": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"name",
					"synopsis",
					"rating",
					"available"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateMovieRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"synopsis": {
						"type": "string"
					},
					"rating": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"synopsis",
					"rating"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateMovieRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"synopsis": {
						"type": "string"
					},
					"rating": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "objective-rental-api",
		"version": "1.0.0",
		"description": "API para locação de filmes - Objective",
		"contact": {}
	},
	"paths": {
		"/api/all": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MovieRepositoryDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Lista todos os filmes disponíveis para locação.",
				"summary": "Listar filmes disponíveis",
				"tags": [
					"Catalog"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/movies": {
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MovieRepositoryDTO"
								}
							}
						}
					}
				},
				"description": "Adiciona um novo filme ao catálogo (Massa de dados).",
				"summary": "Adicionar filme",
				"tags": [
					"Catalog"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateMovieRequest"
							}
						}
					}
				}
			}
		},
		"/api/movies/{id}": {
			"put": {
				"operationId": "Update",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Atualiza os dados de um filme existente.",
				"tags": [
					"Catalog"
				],
				"security": [],
				"parameters": [
					{
						"description": "O UUID do filme",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateMovieRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"204": {
						"description": "No Content"
					}
				},
				"description": "Remove um filme do catálogo.",
				"tags": [
					"Catalog"
				],
				"security": [],
				"parameters": [
					{
						"description": "O UUID do filme",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}